pyproject.toml
[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "protrdx"
version = "0.1.0"
description = "Quant research automation pipeline"
authors = [{name = "Auto Dev"}]
requires-python = ">=3.11"
dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "sqlmodel",
  "pydantic-settings",
  "yfinance",
  "pandas",
  "pandas-ta",
  "matplotlib",
  "apscheduler",
  "openai>=1.0.0",
  "httpx",
  "python-telegram-bot>=20.0",
  "backtesting",
  "pytz",
  "python-dotenv",
  "ruff",
  "mypy",
  "pytest",
  "pytest-asyncio",
  "freezegun"
]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "UP", "B"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = "-q"
asyncio_mode = "auto"
